Move TroopAttribute files to individual module namespace
