+++ WarhammerOldWorld/WarhammerOldWorld/Modules/TroopAttributes/CustomAgentComponents/MoraleAgentComponents/VampireMoraleAgentComponent.cs
