+++ WarhammerOldWorld/WarhammerOldWorld/Modules/TroopAttributes/CustomMissionLogic/VampireAgentMoraleInteractionLogic.cs
